fix linting issue - rename package to avoid underscore
